requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
requests
python-dotenv
pandas
openpyxl
apache-airflow-providers-http
apache-airflow-providers-postgres
apache-airflow-providers-amazon >= 8.0.0